Add simple repeat to movement commands
example:
    * 5l = 5 times to the left
    * 432j = 432 times down

The repeat doesn't work exactly like vim,for example
the left and right movement continues on the next/prev line.

Also if the cursor is at the end/start of the file
and the repeat command can't move it will continue executing
the given number of times doing nothing.