Move to the correct line x possition when moving up and down
Moving between lines the cursor could end in the middle of a multibyte
character.

For example if the cursor is in the third character (single byte)
and we move it a line with double byte characters the cursor will be
placed in the third byte (the middle of the second multibyte char).

To fix it we move the cursor from left to right one character at a time
after moving between lines.