Indent and unindent based on the last char of the previous line
Two new field where added to the editorSyntax struct
indent_chars and unindent_chars helps to identify when a line
should be indented/unindented.

If the previous line has any of the indent_chars as its last
non whitespace character, the next line will be indented.

If the current line has any of the unindent_chars as its first non
whitepace character, the line will be unindented.